+++ docs/changes.rst
@@ -47,10 +47,15 @@ Released: not yet
 
 * Added the 'state' and 'physical-channel-status' properties to the output
   of the "adapter list" command. (issue #472)
+
 * Added 'short-name' and 'reserved-resources' (only when usage options are used)
   columns to the output of the 'partition list' command. (issue #468)
 
 * Added 'description' column to the output of all list commands. (issue #468)
+
+* Added support for missing property options for the 'partition create'
+  and 'partition update' commands. These commands now support options for
+  all properties of z16 HMCs.
 
 **Cleanup:**